// back/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Cargos = {
  CLIENTE: 'CLIENTE',
  ADMIN: 'ADMIN',
  DEV: 'DEV'
} as const

export type Cargos = (typeof Cargos)[keyof typeof Cargos]


export const TipoPagamento = {
  PIX: 'PIX',
  CARTAO: 'CARTAO',
  DINHEIRO: 'DINHEIRO'
} as const

export type TipoPagamento = (typeof TipoPagamento)[keyof typeof TipoPagamento]


export const StatusAgendamento = {
  PENDENTE: 'PENDENTE',
  CONFIRMADO: 'CONFIRMADO',
  CANCELADO: 'CANCELADO',
  REAGENDADO: 'REAGENDADO'
} as const

export type StatusAgendamento = (typeof StatusAgendamento)[keyof typeof StatusAgendamento]


export const StatusAtendimento = {
  PENDENTE: 'PENDENTE',
  INICIADO: 'INICIADO',
  FINALIZADO: 'FINALIZADO'
} as const

export type StatusAtendimento = (typeof StatusAtendimento)[keyof typeof StatusAtendimento]


export const StatusPagamento = {
  PENDENTE: 'PENDENTE',
  PAGO: 'PAGO'
} as const

export type StatusPagamento = (typeof StatusPagamento)[keyof typeof StatusPagamento]